requirements.txt
numpy
pandas
matplotlib
scikit-learn
python-docx
colorama
pyfiglet
requests 